[Watson W] 添加组内销售报表

+++ src/main/java/com/quyeying/charity/report/dto/DataTableResultDto.java
@@ -14,7 +14,7 @@ public class DataTableResultDto {
 
     private int iTotalRecords;
 
-    private int iTotalDisplayRecords;
+    private long iTotalDisplayRecords;
 
     public Object getData() {
         return data;
@@ -32,11 +32,11 @@ public class DataTableResultDto {
         this.iTotalRecords = iTotalRecords;
     }
 
-    public int getiTotalDisplayRecords() {
+    public long getiTotalDisplayRecords() {
         return iTotalDisplayRecords;
     }
 
-    public void setiTotalDisplayRecords(int iTotalDisplayRecords) {
+    public void setiTotalDisplayRecords(long iTotalDisplayRecords) {
         this.iTotalDisplayRecords = iTotalDisplayRecords;
     }
 }
